backend/requirements.txt
fastapi 
uvicorn 
openai 
wikipedia 
python-dotenv
requests